model loader refactoring: wip
