+++ idealingua/idealingua-transpilers/src/main/scala/com/github/pshirshov/izumi/idealingua/il/loader/LocalDomainProcessor.scala
@@ -3,7 +3,6 @@ package com.github.pshirshov.izumi.idealingua.il.loader
 import java.io.File
 import java.nio.file.{Path, Paths}
 
-import com.github.pshirshov.izumi.fundamentals.platform.files.{IzFiles, IzZip}
 import com.github.pshirshov.izumi.idealingua.il.loader.model.{FSPath, LoadedModel}
 import com.github.pshirshov.izumi.idealingua.il.parser.model.{ParsedDomain, ParsedModel}
 import com.github.pshirshov.izumi.idealingua.model.common.DomainId
@@ -76,7 +75,6 @@ protected[loader] class LocalDomainProcessor(
     }
   }
 
-  // TODO: decopypaste?
   private def toModelResolver(incPath: Path): Option[ParsedModel] = {
     val fpath = FSPath(incPath)
     models.results.find(_.path == fpath)
@@ -84,15 +82,32 @@ protected[loader] class LocalDomainProcessor(
         searchClasspath(incPath)
           .flatMap {
             src =>
-              val parsed = parser.parseModels(Map(FSPath(incPath) -> src))
-              parsed.results.find(_.path == fpath)
+              parser
+                .parseModels(Map(fpath -> src))
+                .results.find(_.path == fpath)
           }
       }
       .map(get)
-
   }
 
-  def get(r: ModelParsingResult): ParsedModel = {
+  private def toDomainResolver(incPath: DomainId): Option[ParsedDomain] = {
+    val asPath = toPath(incPath)
+    val fpath = FSPath(asPath)
+
+    domains.results.find(_.path == fpath)
+      .orElse {
+        searchClasspath(asPath)
+          .flatMap {
+            src =>
+              parser
+                .parseDomains(Map(fpath -> src))
+                .results.find(_.path == asPath)
+          }
+      }
+      .map(get)
+  }
+
+  private def get(r: ModelParsingResult): ParsedModel = {
     r match {
       case ModelParsingResult.Success(_, model) =>
         model
@@ -101,7 +116,7 @@ protected[loader] class LocalDomainProcessor(
     }
   }
 
-  def get(r: DomainParsingResult): ParsedDomain = {
+  private def get(r: DomainParsingResult): ParsedDomain = {
     r match {
       case DomainParsingResult.Success(_, d) =>
         d
@@ -110,13 +125,6 @@ protected[loader] class LocalDomainProcessor(
     }
   }
 
-  private def searchClasspath(incPath: Path) = {
-    val fallback = resolveFromCP(incPath, Some("idealingua"))
-      .orElse(resolveFromCP(incPath, None))
-      .orElse(resolveFromJavaCP(incPath))
-      .orElse(resolveFromJars(incPath))
-    fallback
-  }
 
   private def toPath(id: DomainId): Path = {
     val p = Paths.get(id.toPackage.mkString("/"))
@@ -124,53 +132,14 @@ protected[loader] class LocalDomainProcessor(
   }
 
 
-  private def toDomainResolver(incPath: DomainId): Option[ParsedDomain] = {
-    val asPath = toPath(incPath)
-    val fpath = FSPath(asPath)
-
-    domains.results.find(_.path == fpath)
-      .orElse {
-        searchClasspath(asPath).flatMap {
-          src =>
-            val parsed = parser.parseDomains(Map(fpath -> src))
-            parsed.results.find(_.path == asPath)
-        }
-      }.map(get)
-  }
-
-  private def resolveFromCP(incPath: Path, prefix: Option[String]): Option[String] = {
-    val allCandidates = (Seq(root, root.resolve(toPath(domain.did)).getParent).map(_.toFile) ++ classpath)
-      .filter(_.isDirectory)
-      .flatMap {
-        directory =>
-          val base = prefix match {
-            case None =>
-              directory.toPath
-            case Some(v) =>
-              directory.toPath.resolve(v)
-
-          }
-
-          val candidatePath = base.resolve(incPath)
-          val candidates = Seq(candidatePath)
-          candidates.map(_.toFile)
-      }
-
-    val result = allCandidates
-      .find(f => f.exists() && !f.isDirectory)
-      .map(path => IzFiles.readString(path.toPath))
-    result
-  }
-
-  private def resolveFromJavaCP(incPath: Path): Option[String] = {
-    Option(getClass.getResource(Paths.get("/idealingua/").resolve(incPath).toString))
-      .map {
-        fallback =>
-          IzFiles.readString(new File(fallback.toURI).toPath)
-      }
-  }
-
-  private def resolveFromJars(incPath: Path): Option[String] = {
-    IzZip.findInZips(incPath, classpath)
+  private def searchClasspath(incPath: Path): Option[String] = {
+    new ClasspathFileResolver(root,
+      classpath,
+      "idealingua",
+      Seq(root.resolve(toPath(domain.did)).getParent)
+    )
+      .searchClasspath(incPath)
   }
 }
+
+
